profiler in test

+++ cykdtree/tests/test_parallel_kdtree.py
@@ -6,6 +6,12 @@ try:
     from mpi4py import MPI
 except ImportError:  # pragma: w/o MPI
     MPI = None
+try:
+    from Cython.Compiler.Options import directive_defaults
+except ImportError:
+    # Update to cython
+    from Cython.Compiler.Options import get_directive_defaults
+    directive_defaults = get_directive_defaults()
 import cykdtree
 from cykdtree.tests import parametrize, make_points, make_points_neighbors
 from cykdtree.tests.test_parallel_utils import MPITest
@@ -27,8 +33,9 @@ def test_spawn_parallel(nproc=3, npts=20, ndim=2, periodic=False,
                                         left_edge=le, right_edge=re,
                                         periodic=periodic, with_coverage=True,
                                         profile=profile)
-        assert(os.path.isfile(profile))
-        os.remove(profile)
+        if directive_defaults['profile']:
+            assert(os.path.isfile(profile))
+            os.remove(profile)
 
 
 @MPITest(Nproc, periodic=(False, True), ndim=(2,3))
